Extract ReadableByteStreamControllerGetBYOBRequest abstract op
